modify project structure to enable server/api integration tests
