+++ src/main.rs
@@ -1,37 +1,13 @@
+#![allow(unused_imports)]
+#[cfg(feature = "ssr")]
+use actix_web::main;
+
 #[cfg(feature = "ssr")]
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
-    use actix_files::Files;
-    use actix_web::*;
-    use btc_self_custody::app::*;
-    use btc_self_custody::server::health_check::health_check;
-    use leptos::*;
-    use leptos_actix::{generate_route_list, LeptosRoutes};
+    use btc_self_custody::run;
 
-    let conf = get_configuration(None).await.unwrap();
-    let addr = conf.leptos_options.site_addr;
-    // Generate the list of routes in your Leptos App
-    let routes = generate_route_list(|| view! {<App/> });
-    logging::log!("Server listening on: {}", &addr);
-
-    HttpServer::new(move || {
-        let leptos_options = &conf.leptos_options;
-        let site_root = &leptos_options.site_root;
-
-        App::new()
-            .route("/api/{tail:.*}", leptos_actix::handle_server_fns())
-            .route("/health_check", web::get().to(health_check))
-            .leptos_routes(
-                leptos_options.to_owned(),
-                routes.to_owned(),
-                || view! {<App/> },
-            )
-            .service(Files::new("/", site_root))
-        //.wrap(middleware::Compress::default())
-    })
-    .bind(&addr)?
-    .run()
-    .await
+    run().await?.await
 }
 
 #[cfg(not(feature = "ssr"))]
